feat: photo capture modal
